melhora código de rotas e adsense

+++ app/src/main/java/com/example/rebalancear/data/adapters/WalletAssetAdapter.kt
@@ -1,7 +1,7 @@
 package com.example.rebalancear.data.adapters
 
 import com.example.rebalancear.data.models.WalletAssetModel
-import com.example.rebalancear.data.room.entities.WalletAssetRoomEntity
+import com.example.app.data.room.entities.WalletAssetRoomEntity
 import com.example.rebalancear.domain.entities.WalletAsset
 import javax.inject.Inject
 
@@ -12,8 +12,8 @@ class WalletAssetAdapter @Inject constructor() {
     ): WalletAsset {
         return WalletAsset(
             code = userAssetInfo.code,
-            units = userAssetInfo.units,
-            percentGoal = userAssetInfo.goal,
+            units = userAssetInfo.units.toDouble(),
+            percentGoal = userAssetInfo.goal.toDouble(),
             unitPrice = assetModel.unitPrice,
             LPA = assetModel.LPA,
             VPA = assetModel.VPA
